Crate veb

A predecessor data structure based on van Emde Boas trees.

This is implemented with hash maps, so new() is $O(\log\log n)$.

Example

let mut veb = VebSet::new(1000); // capacity
veb.insert(42);
assert!(veb.contains(42));
veb.remove(42);
assert!(!veb.contains(42));

let veb: VebSet = vec![12, 34, 56, 78].into_iter().collect();
assert_eq!(veb.collect(), vec![12, 34, 56, 78]);
assert_eq!(veb.min(), Some(12));
assert_eq!(veb.max(), Some(78));
assert_eq!(veb.succ(34), Some(56));
assert_eq!(veb.pred(34), Some(12));
assert_eq!(veb.contains(34), true);
assert_eq!(veb.contains(35), false);
assert_eq!(veb.len(), 4);

assert_eq!(veb.collect(), vec![12, 34, 56, 78]); // Useful for debugging

Operations

NOTE: min, max, succ, and pred return None if the set is empty.

Operation	Time Complexity	Explanation

insert(x)	$O(\log\log n)$	$S \leftarrow S \cup {x}$

remove(x)	$O(\log\log n)$	$S \leftarrow S \setminus {x}$

contains(x)	$O(\log\log n)$	$x \in S$

min()	$O(1)$	$\min\left( S \right)$

max()	$O(1)$	$\max\left( S \right)$

succ(x)	$O(\log\log n)$	$\min\left( x^△ \right)$

succ_eq(x)	$O(\log\log n)$	$\min\left( x^▲ \right)$

pred(x)	$O(\log\log n)$	$\max\left( x^▽ \right)$

pred_eq(x)	$O(\log\log n)$	$\max\left( x^▲ \right)$

len()	$O(1)$	$|S|$

is_empty()	$O(1)$	$S = \emptyset$

collect()	$O(|S|\log\log n)$	Convert to a Vec

Structs

VebMap

A van Emde Boas tree-based map. The map is implemented as a van Emde Boas tree with a hash map.

Enums

VebSet

A van Emde Boas tree.