Struct veb::VebMap

pub struct VebMap<V> { /* private fields */ }

A van Emde Boas tree-based map. The map is implemented as a van Emde Boas tree with a hash map.

Example

use veb::VebMap;
let mut veb = VebMap::from_iter(vec![(42, "foo"), (43, "bar")]);
assert_eq!(veb.get(42), Some(&"foo"));
assert_eq!(veb.get(43), Some(&"bar"));
assert_eq!(veb.get(44), None);

assert_eq!(veb.min(), Some((42, &"foo")));
assert_eq!(veb.min_key(), Some(42));
assert_eq!(veb.min_value(), Some(&"foo"));

Implementations

impl<V> VebMap<V>

pub fn new(n: usize) -> Self

Creates a new van Emde Boas tree-based map with the given capacity.

Example

let veb = VebMap::<()>::new(1000);

pub fn insert(&mut self, i: usize, v: V) -> Option<V>

Inserts an element into the map. Returns the previous value if the key was already present.

Example

let mut veb = VebMap::new(1000);
assert_eq!(veb.insert(42, "foo"), None);
assert_eq!(veb.insert(42, "bar"), Some("foo"));

pub fn remove(&mut self, i: usize) -> Option<V>

Returns the value at $i$.

Example

let mut veb = VebMap::new(1000);
veb.insert(42, "foo");
assert_eq!(veb.remove(42), Some("foo"));
assert_eq!(veb.remove(42), None);

pub fn get(&self, i: usize) -> Option<&V>

Returns the value corresponding to the key.

Example

let mut veb = VebMap::new(1000);
veb.insert(42, "foo");
assert_eq!(veb.get(42), Some(&"foo"));
assert_eq!(veb.get(43), None);

pub fn get_mut(&mut self, i: usize) -> Option<&mut V>

Returns a mutable reference to the value corresponding to the key.

Example

let mut veb = VebMap::new(1000);
veb.insert(42, "foo");
assert_eq!(veb.get_mut(42), Some(&mut "foo"));
assert_eq!(veb.get_mut(43), None);

pub fn min_key(&self) -> Option<usize>

Returns the key $\min \left( S \right)$.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.min(), None);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.min(), Some((42, &"foo")));

pub fn min_value(&self) -> Option<&V>

Returns the value at $\min \left( S \right)$.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.min_value(), None);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.min_value(), Some(&"foo"));

pub fn min(&self) -> Option<(usize, &V)>

Returns the entry at $\min \left( S \right)$.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.min(), None);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.min(), Some((42, &"foo")));

pub fn max_key(&self) -> Option<usize>

Returns the key $\max \left( S \right)$.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.max(), None);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.max(), Some((42, &"foo")));

pub fn max_value(&self) -> Option<&V>

Returns the value at $\max \left( S \right)$.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.max_value(), None);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.max_value(), Some(&"foo"));

pub fn max(&self) -> Option<(usize, &V)>

Returns the entry at $\max \left( S \right)$.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.max(), None);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.max(), Some((42, &"foo")));

pub fn succ_key(&self, i: usize) -> Option<usize>

Returns the key $\min \left (i^△ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.succ_key(34), Some(56));
assert_eq!(veb.succ_value(34), Some(&"baz"));
assert_eq!(veb.succ(34), Some((56, &"baz")));
assert_eq!(veb.succ(78), None);

pub fn succ_value(&self, i: usize) -> Option<&V>

Returns the value at $\min \left (i^△ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.succ_value(34), Some(&"baz"));
assert_eq!(veb.succ_value(78), None);

pub fn succ(&self, i: usize) -> Option<(usize, &V)>

Returns the entry at $\min \left (i^△ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.succ(34), Some((56, &"baz")));
assert_eq!(veb.succ(78), None);

pub fn succ_eq_key(&self, i: usize) -> Option<usize>

Returns the key $\min \left (i^▲ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.succ_eq_key(34), Some(34));
assert_eq!(veb.succ_eq_key(35), Some(56));

pub fn succ_eq_value(&self, i: usize) -> Option<&V>

Returns the value at $\min \left (i^▲ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.succ_eq_value(34), Some(&"bar"));
assert_eq!(veb.succ_eq_value(35), Some(&"baz"));

pub fn succ_eq(&self, i: usize) -> Option<(usize, &V)>

Returns the entry at $\min \left (i^▲ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.succ_eq(34), Some((34, &"bar")));
assert_eq!(veb.succ_eq(35), Some((56, &"baz")));

pub fn pred_key(&self, i: usize) -> Option<usize>

Returns the key $\max \left (i^▽ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.pred_key(34), Some(12));
assert_eq!(veb.pred_value(34), Some(&"foo"));
assert_eq!(veb.pred(34), Some((12, &"foo")));
assert_eq!(veb.pred(12), None);

pub fn pred_value(&self, i: usize) -> Option<&V>

Returns the value at $\max \left (i^▽ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.pred_value(34), Some(&"foo"));
assert_eq!(veb.pred_value(12), None);

pub fn pred(&self, i: usize) -> Option<(usize, &V)>

Returns the entry at $\max \left (i^▽ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.pred(34), Some((12, &"foo")));
assert_eq!(veb.pred(12), None);

pub fn pred_eq_key(&self, i: usize) -> Option<usize>

Returns the key $\max \left (i^▲ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.pred_eq_key(34), Some(34));
assert_eq!(veb.pred_eq_key(33), Some(12));

pub fn pred_eq_value(&self, i: usize) -> Option<&V>

Returns the value at $\max \left (i^▲ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.pred_eq_value(34), Some(&"bar"));
assert_eq!(veb.pred_eq_value(33), Some(&"foo"));

pub fn pred_eq(&self, i: usize) -> Option<(usize, &V)>

Returns the entry at $\max \left (i^▲ \right)$.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.pred_eq(34), Some((34, &"bar")));
assert_eq!(veb.pred_eq(33), Some((12, &"foo")));

pub fn len(&self) -> usize

Returns $|S|$.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.len(), 0);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the map is empty.

Example

let veb = VebMap::<()>::from_iter(vec![]);
assert_eq!(veb.is_empty(), true);
let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.is_empty(), false);

pub fn contains_key(&self, i: usize) -> bool

Returns true if the map contains the given key.

Example

let veb = VebMap::from_iter(vec![(42, "foo")]);
assert_eq!(veb.contains_key(42), true);
assert_eq!(veb.contains_key(43), false);

pub fn collect(&self) -> Vec<(usize, &V)>

Returns the elements in the map in ascending order. The elements are collected into a Vec.

Example

let veb = VebMap::from_iter(vec![(12, "foo"), (34, "bar"), (56, "baz"), (78, "qux")]);
assert_eq!(veb.collect(), vec![
    (12, &"foo"),
    (34, &"bar"),
    (56, &"baz"),
    (78, &"qux")
]);