Enum veb::VebSet

pub enum VebSet {
    Internal {
        min: usize,
        max: usize,
        len: usize,
        csize: usize,
        summary: Box<VebSet>,
        chunks: HashMap<usize, VebSet>,
    },
    Leaf(u64),
}

A van Emde Boas tree.

Variants

Internal

Fields

min: usize

max: usize

len: usize

csize: usize

summary: Box<VebSet>

chunks: HashMap<usize, VebSet>

Leaf(u64)

Implementations

impl VebSet

pub fn new(n: usize) -> Self

Creates a new van Emde Boas tree with the given capacity.

Example

use veb::VebSet;
let veb = VebSet::new(1_000_000);

pub fn min(&self) -> Option<usize>

Returns the minimum element in the set. Returns None if the set is empty.

Example

assert_eq!(VebSet::from_iter(vec![]).min(), None);
assert_eq!(VebSet::from_iter(vec![42]).min(), Some(42));
assert_eq!(VebSet::from_iter(vec![42, 43]).min(), Some(42));

pub fn max(&self) -> Option<usize>

Returns the maximum element in the set. Returns None if the set is empty.

Example

assert_eq!(VebSet::from_iter(vec![]).max(), None);
assert_eq!(VebSet::from_iter(vec![42]).max(), Some(42));
assert_eq!(VebSet::from_iter(vec![42, 43]).max(), Some(43));

pub fn len(&self) -> usize

Returns the number of elements in the set.

Example

use veb::VebSet;
assert_eq!(VebSet::from_iter(vec![]).len(), 0);
assert_eq!(VebSet::from_iter(vec![42]).len(), 1);
assert_eq!(VebSet::from_iter(vec![42, 43]).len(), 2);

pub fn is_empty(&self) -> bool

Returns true if the set is empty. Returns false if the set is not empty. Equivalent to self.len() == 0.

Example

assert_eq!(VebSet::from_iter(vec![]).is_empty(), true);
assert_eq!(VebSet::from_iter(vec![42]).is_empty(), false);

pub fn insert(&mut self, i: usize) -> bool

Inserts an element into the set. Returns true if the element was not already present. Returns false if the element was already present.

Example

let mut veb = VebSet::new(1_000_000);
assert_eq!(veb.insert(42), true);
assert_eq!(veb.insert(42), false);

pub fn remove(&mut self, i: usize) -> bool

Removes an element from the set. Returns true if the element was present. Returns false if the element was not present.

Example

let mut veb = VebSet::new(1_000_000);
veb.insert(42);
assert_eq!(veb.remove(42), true);
assert_eq!(veb.remove(42), false);

pub fn succ(&self, i: usize) -> Option<usize>

Returns the successor of the given element. Returns None if the given element is the maximum element.

Example

let veb: VebSet = vec![12, 34, 56, 78].into_iter().collect();
assert_eq!(veb.succ(34), Some(56));
assert_eq!(veb.succ(78), None);

pub fn pred(&self, i: usize) -> Option<usize>

Returns the predecessor of the given element. Returns None if the given element is the minimum element.

Example

let veb: VebSet = vec![12, 34, 56, 78].into_iter().collect();
assert_eq!(veb.pred(34), Some(12));
assert_eq!(veb.pred(12), None);

pub fn contains(&self, i: usize) -> bool

Returns true if the set contains the given element. Returns false if the set does not contain the given element.

Example

let veb: VebSet = vec![12, 34, 56, 78].into_iter().collect();
assert_eq!(veb.contains(34), true);
assert_eq!(veb.contains(35), false);

pub fn collect(&self) -> Vec<usize>

Returns the elements in the set in ascending order. The elements are collected into a Vec.